+++ R/generics.R
@@ -27,6 +27,11 @@ setGeneric('spatShift', function(x, ...) standardGeneric('spatShift'))
 setGeneric('overlaps', function(x, ...) standardGeneric('overlaps'))
 
 
+# Object creation ####
+setGeneric('createGiottoPoints', function(x, ...) standardGeneric('createGiottoPoints'))
+setGeneric('createGiottoPolygon', function(x, ...) standardGeneric('createGiottoPolygon'))
+
+
 # Object Characteristics ####
 
 ## nrow() S4 generic ####
@@ -69,4 +74,7 @@ setGeneric('setGiotto', function(gobject, x, ...) standardGeneric('setGiotto'))
 
 # coerce ####
 setGeneric('as.sf', function(x, ...) standardGeneric('as.sf'))
+setGeneric('as.sp', function(x, ...) standardGeneric('as.sp'))
+setGeneric('as.stars', function(x, ...) standardGeneric('as.stars'))
+setGeneric('as.terra', function(x, ...) standardGeneric('as.terra'))