+++ NEWS.md
@@ -4,11 +4,12 @@
 - Giotto subsetting rework
 
 ## Added
-- Added `ext<-()` method for `giottoPolygon`, `giottoPoints`
-- Added `crop()` method for `giottoLargeImage`, `giottoPoints`
 - Added `terraVectData` as virtual parent class for `giottoPolygon` and `giottoPoints` classes
 - Added `$` and `$<-` methods for `terraVectData`
-- Added `[` subsetting for `giottoPoints` and `giottoPolygon`
+- Added `ext<-()` method for `giottoPolygon`, `giottoPoints`
+- Added `crop()` method for `giottoLargeImage`, `giottoPoints`
+- Added `[` subsetting for `giottoPoints` and `giottoPolygon` with numerical, logical, and character (by ID)
+- Added `as.sf()` and `as.stars()` converters for `giottoPoints` and `giottoPolygon`
 - Added `setGiotto()` generic
 
 ## Changes
@@ -16,3 +17,4 @@
 - Updated `spatIDs()` and `featIDs()` methods for `giottoPolygon` and `giottoPoints` to allow returning non-unique IDs
 - Added check for `plot()` when `giottoPolygon` or `giottoPoints` objects contain no geometries
 - Added warning for `crop()` when `giottoLargeImage`, `giottoPolygon`, `giottoPoints` objects are being cropped with an extent that does not include any information
+- Changed: Conversion of `createGiottoPoints()` to a generic function